requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
cssbeautifier==1.15.1
cssselect2==0.7.0
Django==5.0.6
django-unfold==0.38.0
djangorestframework==3.15.2
djlint==1.35.2
drf-yasg==1.21.7
EditorConfig==0.12.4
fonttools==4.53.1
html-tag-names==0.1.2
html-void-elements==0.1.0
html5lib==1.1
idna==3.8
inflection==0.5.1
jsbeautifier==1.15.1
json5==0.9.25
lxml==5.3.0
oscrypto==1.3.0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
pycparser==2.22
pydyf==0.11.0
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
pypdf==4.3.1
pyphen==0.16.0
pypng==0.20220715.0
python-bidi==0.6.0
pytz==2024.2
PyYAML==6.0.2
qrcode==7.4.2
regex==2024.7.24
reportlab==4.2.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
svglib==1.5.1
tinycss2==1.3.0
tomli==2.0.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.2.2
webencodings==0.5.1
xhtml2pdf==0.2.16
zopfli==0.2.3
